bom與product新增修改刪除

+++ src/main/java/com/example/demo/products/ProductdRepository.java
@@ -2,6 +2,6 @@ package com.example.demo.products;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductdRepository extends JpaRepository<Products, Integer> {
+public interface ProductdRepository extends JpaRepository<Products, Long> {
 
 }
